\name{duncan}
\alias{duncan}

\docType{data}
\title{Duncan dataset}
\description{
At 4 points in time the objects (n = 1204 adolescents) were asked to rate cigarette,
marijuana, and alcohol consumption on a 5-point scale.
}
\usage{galo}
\format{Data frame with marijuana (POT), cigarette (CIG), and alcohol (ALC) consumption. 


Category labels: 

1 ... never consumed

2 ... previous but no use over the last 6 months

3 ... current use of less than 4 times a month

4 ... current use of between 4 and 29 times a month

5 ... current use of 30 or more times a month
}

\references{
Duncan, S. C., Duncan, T. E., and Hops, H. (1998). Progressions of alcohol, cigarette,
and marijuana use in adolescence. Journal of Bahavioral Medicine, 21, 375-388.
}


\examples{
   data(duncan)
   duncan
}
\keyword{datasets}
