\name{wurzer}
\alias{wurzer}

\docType{data}
\title{Internet terminals}
\description{
The dataset is about the use of public Internet terminals. For this package we extracted a subset of 8 items. 
}
\usage{wurzer}
\format{A data frame (n = 215) with the following items:

Do you know at least one place where you can finnd such a terminal? (yes/no)

Have you already used such a terminal? (yes/no)

How often do you use the Internet on each of the following locations: home, work, cafe,
terminal, cellphone? (5-point scales; see below)

Which of the following descriptions fits you best? (I'm here on vacation/I am from here/I'm here on business travel)

The 5-point items we have the following categories: daily (1), almost daily (2), several times a
week (3), several times a month (3), once a month (4), less frequently (5). 

}

\references{
Wurzer, M. (2006). An Application of Configural Frequency Analysis: Evaluation of the Usage of Internet Terminals. Master's thesis, University of Vienna, Austria.
}


\examples{
   data(wurzer)
   wurzer
}
\keyword{datasets}
